[COI-5] /my-feed-status/ feeds 앱으로 이동
