--- users/viewsets.py
@@ -1,15 +1,0 @@
-
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.viewsets import ModelViewSet
-
-from .serializers import UserSerializer
-from .models import User
-
-
-# class UserViewSet(ModelViewSet):
-
-#     queryset = User.objects.all()
-#     serializer_class = UserSerializer
-#     permission_classes = [IsAuthenticated]
-
-
